MATLAB Code files- Optimize using WOA
MATLAB Code files to optimize the matrices Q and R of the COLQR Control algorithm; using Whale Optimization Algorithm (WOA). The response curves of the effect of optimizing the matrices using WOA when compared to uncontrolled response can be obtained from the simulink models. Corresponding response curves are also uploaded for quick reference.