Comparison between WOA and RLGA
Simulink models and other related files uploaded. These models help to obtain a comparison of results (displacements and velocities) when the force exerted by the MR damper is controlled using COLQR (wherein the matrices Q and R are controlled using WOA and RLGA). Comparisons can be drawn on the basis of displacements and velocities resulting in each floor in each methodology of control optimization.